Fix material file loading.
A 'long' is not guaranteed to be 4 bytes (in fact it is 8 on my
machine and compiler), but the material file stores a 4-byte
integer. So change the 'long' to 'int32_t'. Unfortunately, I don't
believe there are portable floating point types with a guaranteed bit
width, but it's unlikely that 'float' will not be 32 bits.

+++ src/kofola/audio_plug-in.h
@@ -1,6 +1,8 @@
 #ifndef _AUDIOPLUGIN_
 #define _AUDIOPLUGIN_
 
+#include <stdint.h>
+
 #ifdef __cplusplus
 extern "C"
 {
@@ -13,7 +15,7 @@ extern "C"
 
   typedef struct
   {
-    long Obstruction;
+    int32_t Obstruction;
     float ObstructionLF;
   } MATERIAL_LIST_ITEM;
